Synced code relating to two-fluid thermodynamics:
Now any data file can be run with either plasma_use_p=true (original
default), in which T_e is calculated assuming p_i + p_e = p from the
simulation, or plasma_use_p=false, in which T_e is calculated assuming
u_i + u_e = u from the simulation. The latter case requires knowing the
adiabatic indices for the species separately.

ipole uses the latter formula, though also assumes in the T_e
calculation that n_e/n_i = 1 and mu = 1/2. Here, these parameters always
have the values given by the user in the input file.

The code attempts to be sensible regarding input adiabatic indices. When
they are not needed at all, they are generally ignored, with warnings
issued if they are provided. In the case of Athena++ files, values for
plasma_gamma, when not needed, are silently ignored. Required values
will result in an error if not provided. Sometimes the values are
optional -- the file format must have them, and a warning is issued if
the user provides values and those values do not match the file values.
In some cases the file format might provide them, in which case there is
a warning if inconsistency is found, and there is an error of no source
can be found.